test(cli-utils): refactor spark generate test
